chore: addon_data and settings
