+++ src/gui/element.rs
@@ -30,1154 +30,6 @@ use {
 pub static DEFAULT_FONT_SIZE: u16 = 14;
 pub static DEFAULT_PADDING: u16 = 10;
 
-/// Container for settings.
-pub fn settings_container<'a, 'b>(
-    color_palette: ColorPalette,
-    scrollable_state: &'a mut scrollable::State,
-    directory_button_state: &'a mut button::State,
-    config: &Config,
-    theme_state: &'a mut ThemeState,
-    scale_state: &'a mut ScaleState,
-    backup_state: &'a mut BackupState,
-    column_settings: &'a mut ColumnSettings,
-    column_config: &'b [(ColumnKey, Length, bool)],
-    catalog_column_settings: &'a mut CatalogColumnSettings,
-    catalog_column_config: &'b [(CatalogColumnKey, Length, bool)],
-    open_config_dir_button_state: &'a mut button::State,
-    self_update_channel_state: &'a mut SelfUpdateChannelState,
-) -> Container<'a, Message> {
-    let mut scrollable = Scrollable::new(scrollable_state)
-        .spacing(1)
-        .height(Length::FillPortion(1))
-        .style(style::Scrollable(color_palette));
-
-    // Title for the World of Warcraft directory selection.
-    let directory_info_text = Text::new("World of Warcraft directory").size(DEFAULT_FONT_SIZE);
-    let direction_info_text_container =
-        Container::new(directory_info_text).style(style::BrightBackgroundContainer(color_palette));
-
-    // Directory button for World of Warcraft directory selection.
-    let directory_button_title_container =
-        Container::new(Text::new("Select Directory").size(DEFAULT_FONT_SIZE))
-            .width(Length::FillPortion(1))
-            .center_x()
-            .align_x(Align::Center);
-
-    let directory_button: Element<Interaction> =
-        Button::new(directory_button_state, directory_button_title_container)
-            .width(Length::Units(120))
-            .style(style::DefaultBoxedButton(color_palette))
-            .on_press(Interaction::SelectDirectory(DirectoryType::Wow))
-            .into();
-
-    // Directory text, written next to directory button to let the user
-    // know what has been selected..
-    let path_str = config
-        .wow
-        .directory
-        .as_ref()
-        .and_then(|p| p.to_str())
-        .unwrap_or("No directory is set");
-    let directory_data_text = Text::new(path_str)
-        .size(14)
-        .vertical_alignment(VerticalAlignment::Center);
-    let directory_data_text_container = Container::new(directory_data_text)
-        .height(Length::Units(25))
-        .center_y()
-        .style(style::NormalBackgroundContainer(color_palette));
-
-    // Data row for the World of Warcraft directory selection.
-    let directory_data_row = Row::new()
-        .push(directory_button.map(Message::Interaction))
-        .push(Space::new(Length::Units(DEFAULT_PADDING), Length::Units(0)))
-        .push(directory_data_text_container);
-
-    scrollable = scrollable
-        .push(direction_info_text_container)
-        .push(Space::new(Length::Units(0), Length::Units(5)))
-        .push(directory_data_row);
-
-    let theme_column = {
-        let title_container = Container::new(Text::new("Theme").size(DEFAULT_FONT_SIZE))
-            .style(style::NormalBackgroundContainer(color_palette));
-
-        let theme_names = theme_state
-            .themes
-            .iter()
-            .cloned()
-            .map(|(name, _)| name)
-            .collect::<Vec<_>>();
-        let theme_pick_list = PickList::new(
-            &mut theme_state.pick_list_state,
-            theme_names,
-            Some(theme_state.current_theme_name.clone()),
-            Message::ThemeSelected,
-        )
-        .text_size(14)
-        .width(Length::Units(120))
-        .style(style::PickList(color_palette));
-
-        // Data row for theme picker list.
-        let theme_data_row = Row::new().push(theme_pick_list);
-
-        Column::new()
-            .push(title_container)
-            .push(Space::new(Length::Units(0), Length::Units(5)))
-            .push(theme_data_row)
-    };
-
-    // Scale buttons for application scale factoring.
-    let scale_column = {
-        let title_container = Container::new(Text::new("Scale").size(DEFAULT_FONT_SIZE))
-            .style(style::NormalBackgroundContainer(color_palette));
-        let scale_title_row = Row::new().push(title_container);
-
-        let scale_down_button: Element<Interaction> = Button::new(
-            &mut scale_state.down_btn_state,
-            Text::new("  -  ").size(DEFAULT_FONT_SIZE),
-        )
-        .style(style::DefaultBoxedButton(color_palette))
-        .on_press(Interaction::ScaleDown)
-        .into();
-
-        let scale_up_button: Element<Interaction> = Button::new(
-            &mut scale_state.up_btn_state,
-            Text::new("  +  ").size(DEFAULT_FONT_SIZE),
-        )
-        .style(style::DefaultBoxedButton(color_palette))
-        .on_press(Interaction::ScaleUp)
-        .into();
-
-        let current_scale_text = Text::new(format!("  {:.2}  ", scale_state.scale))
-            .size(DEFAULT_FONT_SIZE)
-            .vertical_alignment(VerticalAlignment::Center);
-        let current_scale_container = Container::new(current_scale_text)
-            .height(Length::Units(25))
-            .center_y()
-            .style(style::BrightBackgroundContainer(color_palette));
-
-        // Data row for the World of Warcraft directory selection.
-        let scale_buttons_row = Row::new()
-            .push(scale_down_button.map(Message::Interaction))
-            .push(current_scale_container)
-            .push(scale_up_button.map(Message::Interaction));
-
-        Column::new()
-            .push(scale_title_row)
-            .push(Space::new(Length::Units(0), Length::Units(5)))
-            .push(scale_buttons_row)
-    };
-
-    let (backup_title_row, backup_directory_row, backup_now_row) = {
-        // Title for the Backup section.
-        let backup_title_text = Text::new("Backup").size(DEFAULT_FONT_SIZE);
-        let backup_title_text_container = Container::new(backup_title_text)
-            .style(style::BrightBackgroundContainer(color_palette));
-
-        let addon_folder_checkbox: Element<_> = Container::new(
-            Checkbox::new(config.backup_addons, "AddOns", move |is_checked| {
-                Interaction::ToggleBackupFolder(is_checked, BackupFolderKind::AddOns)
-            })
-            .text_size(DEFAULT_FONT_SIZE)
-            .spacing(5)
-            .style(style::DefaultCheckbox(color_palette)),
-        )
-        .style(style::BrightBackgroundContainer(color_palette))
-        .into();
-
-        let wtf_folder_checkbox: Element<_> = Container::new(
-            Checkbox::new(config.backup_wtf, "WTF", move |is_checked| {
-                Interaction::ToggleBackupFolder(is_checked, BackupFolderKind::WTF)
-            })
-            .text_size(DEFAULT_FONT_SIZE)
-            .spacing(5)
-            .style(style::DefaultCheckbox(color_palette)),
-        )
-        .style(style::BrightBackgroundContainer(color_palette))
-        .into();
-
-        // Directory button for Backup directory selection.
-        let directory_button_title_container =
-            Container::new(Text::new("Select Directory").size(DEFAULT_FONT_SIZE))
-                .width(Length::FillPortion(1))
-                .center_x()
-                .align_x(Align::Center);
-        let directory_button: Element<Interaction> = Button::new(
-            &mut backup_state.directory_btn_state,
-            directory_button_title_container,
-        )
-        .width(Length::Units(120))
-        .style(style::DefaultBoxedButton(color_palette))
-        .on_press(Interaction::SelectDirectory(DirectoryType::Backup))
-        .into();
-
-        // Directory text, written next to directory button to let the user
-        // know what has been selected.
-        let path_str = config
-            .backup_directory
-            .as_ref()
-            .and_then(|p| p.to_str())
-            .unwrap_or("No directory is set");
-        let directory_data_text = Text::new(path_str)
-            .size(DEFAULT_FONT_SIZE)
-            .vertical_alignment(VerticalAlignment::Center);
-        let directory_data_text_container = Container::new(directory_data_text)
-            .height(Length::Units(25))
-            .center_y()
-            .style(style::NormalBackgroundContainer(color_palette));
-
-        // Data row for the Backup directory selection.
-        let backup_directory_row = Row::new()
-            .align_items(Align::Center)
-            .height(Length::Units(26))
-            .push(addon_folder_checkbox.map(Message::Interaction))
-            .push(Space::new(Length::Units(DEFAULT_PADDING), Length::Units(0)))
-            .push(wtf_folder_checkbox.map(Message::Interaction))
-            .push(Space::new(Length::Units(DEFAULT_PADDING), Length::Units(0)))
-            .push(directory_button.map(Message::Interaction))
-            .push(Space::new(Length::Units(DEFAULT_PADDING), Length::Units(0)))
-            .push(directory_data_text_container);
-
-        // Row to show actual backup button along with info about the latest
-        // backup date/time. Will give a description of what Backup is when no
-        // directory is chosen
-        let mut backup_now_row = Row::new();
-
-        // Show button / last backup info if directory is shown, otherwise
-        // show description about the backup process
-        if config.backup_directory.is_some() {
-            let backup_button_title_container =
-                Container::new(Text::new("Backup Now").size(DEFAULT_FONT_SIZE))
-                    .width(Length::FillPortion(1))
-                    .center_x()
-                    .align_x(Align::Center);
-            let mut backup_button = Button::new(
-                &mut backup_state.backup_now_btn_state,
-                backup_button_title_container,
-            )
-            .width(Length::Units(120))
-            .style(style::DefaultBoxedButton(color_palette));
-
-            // Only show button as clickable if it's not currently backing up and
-            // the wow folder is chosen and at least one of the folders is selected
-            // for backup
-            if !backup_state.backing_up
-                && config.wow.directory.is_some()
-                && (config.backup_addons || config.backup_wtf)
-            {
-                backup_button = backup_button.on_press(Interaction::Backup);
-            }
-
-            let backup_status_text = if backup_state.backing_up {
-                Text::new("Backing up...")
-                    .size(DEFAULT_FONT_SIZE)
-                    .vertical_alignment(VerticalAlignment::Center)
-            } else {
-                let as_of = backup_state
-                    .last_backup
-                    .map(|d| d.format("%Y-%m-%d %H:%M:%S").to_string())
-                    .unwrap_or_else(|| "Never".to_string());
-
-                Text::new(&format!("Last backup: {}", as_of))
-                    .size(DEFAULT_FONT_SIZE)
-                    .vertical_alignment(VerticalAlignment::Center)
-            };
-
-            let backup_status_text_container = Container::new(backup_status_text)
-                .height(Length::Units(25))
-                .center_y()
-                .style(style::NormalBackgroundContainer(color_palette));
-
-            let backup_button: Element<Interaction> = backup_button.into();
-
-            backup_now_row = backup_now_row
-                .push(backup_button.map(Message::Interaction))
-                .push(Space::new(Length::Units(DEFAULT_PADDING), Length::Units(0)))
-                .push(backup_status_text_container);
-        } else {
-            let backup_status_text =
-                Text::new("Back up your AddOns and WTF folder to the chosen directory")
-                    .size(DEFAULT_FONT_SIZE)
-                    .vertical_alignment(VerticalAlignment::Center);
-
-            let backup_status_text_container = Container::new(backup_status_text)
-                .height(Length::Units(25))
-                .center_y()
-                .style(style::NormalBackgroundContainer(color_palette));
-
-            backup_now_row = backup_now_row.push(backup_status_text_container);
-        }
-
-        (
-            backup_title_text_container,
-            backup_directory_row,
-            backup_now_row,
-        )
-    };
-
-    let hide_addons_column = {
-        let hide_ignored_addons = config.hide_ignored_addons;
-        let title = "Hide ignored Addons".to_owned();
-        let checkbox = Checkbox::new(hide_ignored_addons, title, move |is_checked| {
-            Message::Interaction(Interaction::ToggleHideIgnoredAddons(is_checked))
-        })
-        .style(style::DefaultCheckbox(color_palette))
-        .text_size(DEFAULT_FONT_SIZE)
-        .spacing(5);
-        let checkbox_container =
-            Container::new(checkbox).style(style::BrightBackgroundContainer(color_palette));
-        Column::new().push(checkbox_container)
-    };
-
-    let config_column = {
-        let config_dir = ajour_core::fs::config_dir();
-        let config_dir_string = config_dir.as_path().display().to_string();
-
-        let open_config_button_title_container =
-            Container::new(Text::new("Open data directory").size(DEFAULT_FONT_SIZE))
-                .width(Length::Units(150))
-                .center_x()
-                .align_x(Align::Center);
-        let open_config_button: Element<Interaction> = Button::new(
-            open_config_dir_button_state,
-            open_config_button_title_container,
-        )
-        .width(Length::Units(150))
-        .style(style::DefaultBoxedButton(color_palette))
-        .on_press(Interaction::OpenDirectory(config_dir))
-        .into();
-
-        let open_config_description = Text::new(config_dir_string)
-            .size(DEFAULT_FONT_SIZE)
-            .vertical_alignment(VerticalAlignment::Center);
-        let open_config_description_container = Container::new(open_config_description)
-            .height(Length::Units(25))
-            .center_y()
-            .style(style::NormalBackgroundContainer(color_palette));
-
-        let open_config_row = Row::new()
-            .push(open_config_button.map(Message::Interaction))
-            .push(Space::new(Length::Units(DEFAULT_PADDING), Length::Units(0)))
-            .push(open_config_description_container);
-
-        Column::new().push(open_config_row)
-    };
-
-    let ui_title = Text::new("UI").size(DEFAULT_FONT_SIZE);
-    let ui_title_container =
-        Container::new(ui_title).style(style::BrightBackgroundContainer(color_palette));
-
-    let addon_title = Text::new("Addons").size(DEFAULT_FONT_SIZE);
-    let addon_title_container =
-        Container::new(addon_title).style(style::BrightBackgroundContainer(color_palette));
-
-    let config_title = Text::new("Config").size(DEFAULT_FONT_SIZE);
-    let config_title_container =
-        Container::new(config_title).style(style::BrightBackgroundContainer(color_palette));
-
-    let ui_row = Row::new()
-        .push(theme_column)
-        .push(scale_column)
-        .spacing(DEFAULT_PADDING);
-
-    let channel_title = Container::new(Text::new("Self Update Channel").size(DEFAULT_FONT_SIZE))
-        .style(style::BrightBackgroundContainer(color_palette));
-    let channel_picklist: Element<_> = PickList::new(
-        &mut self_update_channel_state.picklist,
-        &self_update_channel_state.options[..],
-        Some(config.self_update_channel),
-        Interaction::PickSelfUpdateChannel,
-    )
-    .text_size(14)
-    .width(Length::Fill)
-    .style(style::PickList(color_palette))
-    .into();
-
-    let channel_container = Container::new(channel_picklist.map(Message::Interaction))
-        .center_y()
-        .width(Length::Units(80))
-        .style(style::NormalForegroundContainer(color_palette));
-
-    scrollable = scrollable
-        .push(Space::new(Length::Units(0), Length::Units(20)))
-        .push(backup_title_row)
-        .push(Space::new(Length::Units(0), Length::Units(5)))
-        .push(backup_now_row)
-        .push(Space::new(Length::Units(0), Length::Units(5)))
-        .push(backup_directory_row)
-        .push(Space::new(Length::Units(0), Length::Units(20)))
-        .push(channel_title)
-        .push(Space::new(Length::Units(0), Length::Units(5)))
-        .push(channel_container)
-        .push(Space::new(Length::Units(0), Length::Units(20)))
-        .push(addon_title_container)
-        .push(Space::new(Length::Units(0), Length::Units(5)))
-        .push(hide_addons_column)
-        .push(Space::new(Length::Units(0), Length::Units(20)))
-        .push(ui_title_container)
-        .push(Space::new(Length::Units(0), Length::Units(5)))
-        .push(ui_row)
-        .push(Space::new(Length::Units(0), Length::Units(20)))
-        .push(config_title_container)
-        .push(Space::new(Length::Units(0), Length::Units(5)))
-        .push(config_column);
-
-    let columns_title_text = Text::new("Columns").size(DEFAULT_FONT_SIZE);
-    let columns_title_text_container =
-        Container::new(columns_title_text).style(style::BrightBackgroundContainer(color_palette));
-    scrollable = scrollable
-        .push(Space::new(Length::Units(0), Length::Units(20)))
-        .push(columns_title_text_container);
-
-    let my_addons_columns_container = {
-        let title_container = Container::new(Text::new("My Addons").size(DEFAULT_FONT_SIZE))
-            .style(style::NormalBackgroundContainer(color_palette));
-        let mut my_addons_column = Column::new()
-            .push(title_container)
-            .push(Space::new(Length::Units(0), Length::Units(DEFAULT_PADDING)));
-
-        // Add each column to scrollable as checkbox + label + up / down buttons
-        let columns_len = column_settings.columns.len();
-        for (idx, column) in column_settings.columns.iter_mut().enumerate() {
-            let is_first = idx == 0;
-            let is_last = idx == columns_len - 1;
-
-            let title = column.key.title();
-
-            let column_key = column.key;
-
-            let is_checked = column_config
-                .iter()
-                .any(|(key, _, hidden)| key == &column.key && !hidden)
-                || column_key == ColumnKey::Title;
-
-            let mut left_button = Button::new(
-                &mut column.up_btn_state,
-                Text::new(" ▲ ").size(11).color(if !is_first {
-                    color_palette.bright.primary
-                } else {
-                    color_palette.normal.primary
-                }),
-            )
-            .style(style::DefaultButton(color_palette));
-            if !is_first {
-                left_button = left_button.on_press(Interaction::MoveColumnLeft(column_key));
-            }
-
-            let mut right_button = Button::new(
-                &mut column.down_btn_state,
-                Text::new(" ▼ ").size(11).color(if !is_last {
-                    color_palette.bright.primary
-                } else {
-                    color_palette.normal.primary
-                }),
-            )
-            .style(style::DefaultButton(color_palette));
-            if !is_last {
-                right_button = right_button.on_press(Interaction::MoveColumnRight(column_key));
-            }
-
-            let left_button: Element<Interaction> = left_button.into();
-            let right_button: Element<Interaction> = right_button.into();
-
-            let left_button_container = Container::new(left_button.map(Message::Interaction))
-                .center_x()
-                .center_y();
-            let right_button_container = Container::new(right_button.map(Message::Interaction))
-                .center_x()
-                .center_y();
-
-            let mut checkbox = Checkbox::new(is_checked, title.clone(), move |is_checked| {
-                Message::Interaction(Interaction::ToggleColumn(is_checked, column_key))
-            })
-            .text_size(DEFAULT_FONT_SIZE)
-            .spacing(5);
-
-            if column_key == ColumnKey::Title {
-                checkbox = checkbox.style(style::AlwaysCheckedCheckbox(color_palette));
-            } else {
-                checkbox = checkbox.style(style::DefaultCheckbox(color_palette));
-            }
-
-            let checkbox_container =
-                Container::new(checkbox).style(style::BrightBackgroundContainer(color_palette));
-
-            let row = Row::new()
-                .align_items(Align::Center)
-                .height(Length::Units(26))
-                .push(left_button_container)
-                .push(right_button_container)
-                .push(checkbox_container);
-
-            my_addons_column = my_addons_column.push(row);
-        }
-
-        Container::new(my_addons_column)
-    };
-
-    let catalog_columns_container = {
-        // Title for the Columns section.
-
-        let title_container = Container::new(Text::new("Catalog").size(DEFAULT_FONT_SIZE))
-            .style(style::NormalBackgroundContainer(color_palette));
-        let mut catalog_column = Column::new()
-            .push(title_container)
-            .push(Space::new(Length::Units(0), Length::Units(DEFAULT_PADDING)));
-
-        // Add each column to scrollable as checkbox + label + up / down buttons
-        let columns_len = catalog_column_settings.columns.len();
-        for (idx, column) in catalog_column_settings.columns.iter_mut().enumerate() {
-            let is_first = idx == 0;
-            let is_last = idx == columns_len - 1;
-
-            let title = column.key.title();
-
-            let column_key = column.key;
-
-            let is_checked = catalog_column_config
-                .iter()
-                .any(|(key, _, hidden)| key == &column.key && !hidden)
-                || column_key == CatalogColumnKey::Title;
-
-            let mut left_button = Button::new(
-                &mut column.up_btn_state,
-                Text::new(" ▲  ").size(11).color(if !is_first {
-                    color_palette.bright.primary
-                } else {
-                    color_palette.normal.primary
-                }),
-            )
-            .style(style::DefaultButton(color_palette));
-            if !is_first {
-                left_button = left_button.on_press(Interaction::MoveCatalogColumnLeft(column_key));
-            }
-
-            let mut right_button = Button::new(
-                &mut column.down_btn_state,
-                Text::new(" ▼  ").size(11).color(if !is_last {
-                    color_palette.bright.primary
-                } else {
-                    color_palette.normal.primary
-                }),
-            )
-            .style(style::DefaultButton(color_palette));
-            if !is_last {
-                right_button =
-                    right_button.on_press(Interaction::MoveCatalogColumnRight(column_key));
-            }
-
-            let left_button: Element<Interaction> = left_button.into();
-            let right_button: Element<Interaction> = right_button.into();
-
-            let left_button_container = Container::new(left_button.map(Message::Interaction))
-                .center_x()
-                .center_y();
-            let right_button_container = Container::new(right_button.map(Message::Interaction))
-                .center_x()
-                .center_y();
-
-            let mut checkbox = Checkbox::new(is_checked, title.clone(), move |is_checked| {
-                Message::Interaction(Interaction::ToggleCatalogColumn(is_checked, column_key))
-            })
-            .text_size(DEFAULT_FONT_SIZE)
-            .spacing(5);
-
-            if column_key == CatalogColumnKey::Title {
-                checkbox = checkbox.style(style::AlwaysCheckedCheckbox(color_palette));
-            } else {
-                checkbox = checkbox.style(style::DefaultCheckbox(color_palette));
-            }
-
-            let checkbox_container =
-                Container::new(checkbox).style(style::BrightBackgroundContainer(color_palette));
-
-            let row = Row::new()
-                .align_items(Align::Center)
-                .height(Length::Units(26))
-                .push(left_button_container)
-                .push(right_button_container)
-                .push(checkbox_container);
-
-            catalog_column = catalog_column.push(row);
-        }
-
-        Container::new(catalog_column)
-    };
-
-    let catalog_rows = Row::new()
-        .push(my_addons_columns_container)
-        .push(catalog_columns_container)
-        .spacing(DEFAULT_PADDING);
-
-    scrollable = scrollable
-        .push(Space::new(Length::Units(0), Length::Units(5)))
-        .push(catalog_rows);
-
-    // Colum wrapping all the settings content.
-    scrollable = scrollable.height(Length::Fill).width(Length::Fill);
-
-    let col = Column::new()
-        .push(Space::new(Length::Units(0), Length::Units(10)))
-        .push(scrollable)
-        .push(Space::new(Length::Units(0), Length::Units(20)));
-    let row = Row::new()
-        .push(Space::new(Length::Units(20), Length::Units(0)))
-        .push(col);
-
-    // Returns the final container.
-    Container::new(row)
-        .center_x()
-        .width(Length::Fill)
-        .height(Length::Shrink)
-        .style(style::NormalBackgroundContainer(color_palette))
-}
-
-pub fn addon_data_cell<'a, 'b>(
-    color_palette: ColorPalette,
-    addon: &'a mut Addon,
-    is_addon_expanded: bool,
-    expand_type: &'a ExpandType,
-    column_config: &'b [(ColumnKey, Length, bool)],
-) -> Container<'a, Message> {
-    let default_height = Length::Units(26);
-
-    let mut row_containers = vec![];
-
-    let author = addon.author().map(str::to_string);
-    let game_version = addon.game_version().map(str::to_string);
-    let notes = addon.notes().map(str::to_string);
-    let website_url = addon.website_url().map(str::to_string);
-    let changelog_url = addon.changelog_url().map(str::to_string);
-    let repository_kind = addon.repository_kind();
-
-    // Check if current addon is expanded.
-    let addon_cloned = addon.clone();
-    let version = addon
-        .version()
-        .map(str::to_string)
-        .unwrap_or_else(|| "-".to_string());
-    let release_package = addon_cloned.relevant_release_package();
-    let remote_version = if let Some(package) = &release_package {
-        package.version.clone()
-    } else {
-        String::from("-")
-    };
-    let remote_version = Text::new(remote_version).size(DEFAULT_FONT_SIZE);
-
-    if let Some((idx, width)) = column_config
-        .iter()
-        .enumerate()
-        .filter_map(|(idx, (key, width, hidden))| {
-            if *key == ColumnKey::Title && !hidden {
-                Some((idx, width))
-            } else {
-                None
-            }
-        })
-        .next()
-    {
-        let title = Text::new(addon.title()).size(DEFAULT_FONT_SIZE);
-        let mut title_button = Button::new(&mut addon.details_btn_state, title)
-            .on_press(Interaction::Expand(ExpandType::Details(addon_cloned)));
-
-        if release_package.is_some() {}
-
-        if is_addon_expanded && matches!(expand_type, ExpandType::Details(_)) {
-            title_button = title_button.style(style::SelectedBrightTextButton(color_palette));
-        } else {
-            title_button = title_button.style(style::BrightTextButton(color_palette));
-        }
-
-        let title_button: Element<Interaction> = title_button.into();
-
-        let mut title_row = Row::new()
-            .push(title_button.map(Message::Interaction))
-            .spacing(3)
-            .align_items(Align::Center);
-
-        if addon.release_channel != ReleaseChannel::Stable {
-            let release_channel =
-                Container::new(Text::new(addon.release_channel.to_string()).size(10))
-                    .style(style::ChannelBadge(color_palette))
-                    .padding(3);
-
-            title_row = title_row.push(release_channel);
-        }
-
-        let title_container = Container::new(title_row)
-            .height(default_height)
-            .width(*width)
-            .center_y()
-            .style(style::BrightForegroundContainer(color_palette));
-
-        row_containers.push((idx, title_container));
-    }
-
-    if let Some((idx, width)) = column_config
-        .iter()
-        .enumerate()
-        .filter_map(|(idx, (key, width, hidden))| {
-            if *key == ColumnKey::LocalVersion && !hidden {
-                Some((idx, width))
-            } else {
-                None
-            }
-        })
-        .next()
-    {
-        let installed_version = Text::new(version).size(DEFAULT_FONT_SIZE);
-
-        let installed_version_container = Container::new(installed_version)
-            .padding(5)
-            .height(default_height)
-            .width(*width)
-            .center_y()
-            .style(style::NormalForegroundContainer(color_palette));
-
-        row_containers.push((idx, installed_version_container));
-    }
-
-    if let Some((idx, width)) = column_config
-        .iter()
-        .enumerate()
-        .filter_map(|(idx, (key, width, hidden))| {
-            if *key == ColumnKey::RemoteVersion && !hidden {
-                Some((idx, width))
-            } else {
-                None
-            }
-        })
-        .next()
-    {
-        let remote_version_container = Container::new(remote_version)
-            .padding(5)
-            .height(default_height)
-            .width(*width)
-            .center_y()
-            .style(style::NormalForegroundContainer(color_palette));
-
-        row_containers.push((idx, remote_version_container));
-    }
-
-    if let Some((idx, width)) = column_config
-        .iter()
-        .enumerate()
-        .filter_map(|(idx, (key, width, hidden))| {
-            if *key == ColumnKey::Channel && !hidden {
-                Some((idx, width))
-            } else {
-                None
-            }
-        })
-        .next()
-    {
-        let channel = Text::new(addon.release_channel.to_string()).size(DEFAULT_FONT_SIZE);
-        let channel_container = Container::new(channel)
-            .height(default_height)
-            .width(*width)
-            .center_y()
-            .padding(5)
-            .style(style::NormalForegroundContainer(color_palette));
-
-        row_containers.push((idx, channel_container));
-    }
-
-    if let Some((idx, width)) = column_config
-        .iter()
-        .enumerate()
-        .filter_map(|(idx, (key, width, hidden))| {
-            if *key == ColumnKey::Author && !hidden {
-                Some((idx, width))
-            } else {
-                None
-            }
-        })
-        .next()
-    {
-        let author = Text::new(author.as_deref().unwrap_or("-")).size(DEFAULT_FONT_SIZE);
-        let author_container = Container::new(author)
-            .height(default_height)
-            .width(*width)
-            .center_y()
-            .padding(5)
-            .style(style::NormalForegroundContainer(color_palette));
-
-        row_containers.push((idx, author_container));
-    }
-
-    if let Some((idx, width)) = column_config
-        .iter()
-        .enumerate()
-        .filter_map(|(idx, (key, width, hidden))| {
-            if *key == ColumnKey::GameVersion && !hidden {
-                Some((idx, width))
-            } else {
-                None
-            }
-        })
-        .next()
-    {
-        let game_version =
-            Text::new(game_version.as_deref().unwrap_or("-")).size(DEFAULT_FONT_SIZE);
-        let game_version_container = Container::new(game_version)
-            .height(default_height)
-            .width(*width)
-            .center_y()
-            .padding(5)
-            .style(style::NormalForegroundContainer(color_palette));
-
-        row_containers.push((idx, game_version_container));
-    }
-
-    if let Some((idx, width)) = column_config
-        .iter()
-        .enumerate()
-        .filter_map(|(idx, (key, width, hidden))| {
-            if *key == ColumnKey::DateReleased && !hidden {
-                Some((idx, width))
-            } else {
-                None
-            }
-        })
-        .next()
-    {
-        let release_date_text: String = if let Some(package) = &release_package {
-            let f = timeago::Formatter::new();
-            let now = Local::now();
-
-            if let Some(time) = package.date_time.as_ref() {
-                f.convert_chrono(*time, now)
-            } else {
-                "".to_string()
-            }
-        } else {
-            "-".to_string()
-        };
-        let release_date_text = Text::new(release_date_text).size(DEFAULT_FONT_SIZE);
-        let game_version_container = Container::new(release_date_text)
-            .height(default_height)
-            .width(*width)
-            .center_y()
-            .padding(5)
-            .style(style::NormalForegroundContainer(color_palette));
-
-        row_containers.push((idx, game_version_container));
-    }
-
-    if let Some((idx, width)) = column_config
-        .iter()
-        .enumerate()
-        .filter_map(|(idx, (key, width, hidden))| {
-            if *key == ColumnKey::Source && !hidden {
-                Some((idx, width))
-            } else {
-                None
-            }
-        })
-        .next()
-    {
-        let source_text =
-            repository_kind.map_or_else(|| String::from("Unknown"), |a| a.to_string());
-        let source = Text::new(source_text).size(DEFAULT_FONT_SIZE);
-        let source_container = Container::new(source)
-            .height(default_height)
-            .width(*width)
-            .center_y()
-            .padding(5)
-            .style(style::NormalForegroundContainer(color_palette));
-
-        row_containers.push((idx, source_container));
-    }
-
-    if let Some((idx, width)) = column_config
-        .iter()
-        .enumerate()
-        .filter_map(|(idx, (key, width, hidden))| {
-            if *key == ColumnKey::Status && !hidden {
-                Some((idx, width))
-            } else {
-                None
-            }
-        })
-        .next()
-    {
-        let update_button_container = match &addon.state {
-            AddonState::Idle => Container::new(Text::new("".to_string()).size(DEFAULT_FONT_SIZE))
-                .height(default_height)
-                .width(*width)
-                .center_y()
-                .center_x()
-                .style(style::NormalForegroundContainer(color_palette)),
-            AddonState::Completed => {
-                Container::new(Text::new("Completed".to_string()).size(DEFAULT_FONT_SIZE))
-                    .height(default_height)
-                    .width(*width)
-                    .center_y()
-                    .center_x()
-                    .style(style::NormalForegroundContainer(color_palette))
-            }
-            AddonState::Error(message) => {
-                Container::new(Text::new(message).size(DEFAULT_FONT_SIZE))
-                    .height(default_height)
-                    .width(*width)
-                    .center_y()
-                    .center_x()
-                    .style(style::NormalForegroundContainer(color_palette))
-            }
-            AddonState::Updatable | AddonState::Retry => {
-                let id = addon.primary_folder_id.clone();
-
-                let text = match addon.state {
-                    AddonState::Updatable => "Update",
-                    AddonState::Retry => "Retry",
-                    _ => "",
-                };
-
-                let update_wrapper = Container::new(Text::new(text).size(DEFAULT_FONT_SIZE))
-                    .width(*width)
-                    .center_x()
-                    .align_x(Align::Center);
-                let update_button: Element<Interaction> =
-                    Button::new(&mut addon.update_btn_state, update_wrapper)
-                        .width(Length::FillPortion(1))
-                        .style(style::SecondaryButton(color_palette))
-                        .on_press(Interaction::Update(id))
-                        .into();
-
-                Container::new(update_button.map(Message::Interaction))
-                    .height(default_height)
-                    .width(*width)
-                    .center_y()
-                    .center_x()
-                    .style(style::BrightForegroundContainer(color_palette))
-            }
-            AddonState::Downloading => {
-                Container::new(Text::new("Downloading").size(DEFAULT_FONT_SIZE))
-                    .height(default_height)
-                    .width(*width)
-                    .center_y()
-                    .center_x()
-                    .padding(5)
-                    .style(style::NormalForegroundContainer(color_palette))
-            }
-            AddonState::Unpacking => Container::new(Text::new("Unpacking").size(DEFAULT_FONT_SIZE))
-                .height(default_height)
-                .width(*width)
-                .center_y()
-                .center_x()
-                .padding(5)
-                .style(style::NormalForegroundContainer(color_palette)),
-            AddonState::Fingerprint => Container::new(Text::new("Hashing").size(DEFAULT_FONT_SIZE))
-                .height(default_height)
-                .width(*width)
-                .center_y()
-                .center_x()
-                .padding(5)
-                .style(style::NormalForegroundContainer(color_palette)),
-            AddonState::Ignored => Container::new(Text::new("Ignored").size(DEFAULT_FONT_SIZE))
-                .height(default_height)
-                .width(*width)
-                .center_y()
-                .center_x()
-                .padding(5)
-                .style(style::NormalForegroundContainer(color_palette)),
-            AddonState::Unknown => Container::new(Text::new("").size(DEFAULT_FONT_SIZE))
-                .height(default_height)
-                .width(*width)
-                .center_y()
-                .center_x()
-                .padding(5)
-                .style(style::NormalForegroundContainer(color_palette)),
-        };
-
-        row_containers.push((idx, update_button_container));
-    }
-
-    let left_spacer = Space::new(Length::Units(DEFAULT_PADDING), Length::Units(0));
-    let right_spacer = Space::new(Length::Units(DEFAULT_PADDING + 5), Length::Units(0));
-
-    let mut row = Row::new().push(left_spacer).spacing(1);
-
-    // Sort columns and push them into row
-    row_containers.sort_by(|a, b| a.0.cmp(&b.0));
-    for (_, elem) in row_containers.into_iter() {
-        row = row.push(elem);
-    }
-
-    row = row.push(right_spacer);
-
-    let mut addon_column = Column::new().push(row);
-
-    if is_addon_expanded {
-        if let ExpandType::Details(_) = expand_type {
-            let notes = notes.unwrap_or_else(|| "No description for addon.".to_string());
-            let author = author.unwrap_or_else(|| "-".to_string());
-            let left_spacer = Space::new(Length::Units(DEFAULT_PADDING), Length::Units(0));
-            let space = Space::new(Length::Units(0), Length::Units(DEFAULT_PADDING * 2));
-            let bottom_space = Space::new(Length::Units(0), Length::Units(4));
-            let notes_title_text = Text::new("Summary").size(DEFAULT_FONT_SIZE);
-            let notes_text = Text::new(notes).size(DEFAULT_FONT_SIZE);
-            let author_text = Text::new(author).size(DEFAULT_FONT_SIZE);
-            let author_title_text = Text::new("Author(s)").size(DEFAULT_FONT_SIZE);
-            let author_title_container = Container::new(author_title_text)
-                .style(style::FadedBrightForegroundContainer(color_palette));
-            let notes_title_container = Container::new(notes_title_text)
-                .style(style::FadedBrightForegroundContainer(color_palette));
-
-            let release_date_text: String = if let Some(package) = &release_package {
-                let f = timeago::Formatter::new();
-                let now = Local::now();
-
-                if let Some(time) = package.date_time.as_ref() {
-                    format!("is {}", f.convert_chrono(*time, now))
-                } else {
-                    "".to_string()
-                }
-            } else {
-                "has no avaiable release".to_string()
-            };
-            let release_date_text = Text::new(release_date_text).size(DEFAULT_FONT_SIZE);
-            let release_date_text_container = Container::new(release_date_text)
-                .center_y()
-                .padding(5)
-                .style(style::FadedBrightForegroundContainer(color_palette));
-
-            let release_channel_title = Text::new("Remote release channel").size(DEFAULT_FONT_SIZE);
-            let release_channel_title_container = Container::new(release_channel_title)
-                .style(style::FadedBrightForegroundContainer(color_palette));
-            let release_channel_list = PickList::new(
-                &mut addon.pick_release_channel_state,
-                &ReleaseChannel::ALL[..],
-                Some(addon.release_channel),
-                Message::ReleaseChannelSelected,
-            )
-            .text_size(14)
-            .width(Length::Units(100))
-            .style(style::PickList(color_palette));
-
-            let mut website_button = Button::new(
-                &mut addon.website_btn_state,
-                Text::new("Website").size(DEFAULT_FONT_SIZE),
-            )
-            .style(style::DefaultButton(color_palette));
-
-            if let Some(link) = website_url {
-                website_button = website_button.on_press(Interaction::OpenLink(link));
-            }
-
-            let website_button: Element<Interaction> = website_button.into();
-
-            let mut force_download_button = Button::new(
-                &mut addon.force_btn_state,
-                Text::new("Force update").size(DEFAULT_FONT_SIZE),
-            )
-            .style(style::DefaultButton(color_palette));
-
-            // If we have a release package on addon, enable force update.
-            if release_package.is_some() {
-                force_download_button = force_download_button
-                    .on_press(Interaction::Update(addon.primary_folder_id.clone()));
-            }
-
-            let force_download_button: Element<Interaction> = force_download_button.into();
-
-            let is_ignored = addon.state == AddonState::Ignored;
-            let ignore_button_text = if is_ignored {
-                Text::new("Unignore").size(DEFAULT_FONT_SIZE)
-            } else {
-                Text::new("Ignore").size(DEFAULT_FONT_SIZE)
-            };
-
-            let mut ignore_button = Button::new(&mut addon.ignore_btn_state, ignore_button_text)
-                .on_press(Interaction::Ignore(addon.primary_folder_id.clone()))
-                .style(style::DefaultButton(color_palette));
-
-            if is_ignored {
-                ignore_button =
-                    ignore_button.on_press(Interaction::Unignore(addon.primary_folder_id.clone()));
-            } else {
-                ignore_button =
-                    ignore_button.on_press(Interaction::Ignore(addon.primary_folder_id.clone()));
-            }
-
-            let ignore_button: Element<Interaction> = ignore_button.into();
-
-            let delete_button: Element<Interaction> = Button::new(
-                &mut addon.delete_btn_state,
-                Text::new("Delete").size(DEFAULT_FONT_SIZE),
-            )
-            .on_press(Interaction::Delete(addon.primary_folder_id.clone()))
-            .style(style::DefaultDeleteButton(color_palette))
-            .into();
-
-            let mut changelog_button = Button::new(
-                &mut addon.changelog_btn_state,
-                Text::new("Changelog").size(DEFAULT_FONT_SIZE),
-            )
-            .style(style::DefaultButton(color_palette));
-
-            if let Some(link) = changelog_url {
-                changelog_button = changelog_button.on_press(Interaction::OpenLink(link));
-            }
-
-            let changelog_button: Element<Interaction> = changelog_button.into();
-
-            let test_row = Row::new()
-                .push(release_channel_list)
-                .push(release_date_text_container);
-
-            let button_row = Row::new()
-                .push(Space::new(Length::Fill, Length::Units(0)))
-                .push(website_button.map(Message::Interaction))
-                .push(Space::new(Length::Units(5), Length::Units(0)))
-                .push(changelog_button.map(Message::Interaction))
-                .push(Space::new(Length::Units(5), Length::Units(0)))
-                .push(force_download_button.map(Message::Interaction))
-                .push(Space::new(Length::Units(5), Length::Units(0)))
-                .push(ignore_button.map(Message::Interaction))
-                .push(Space::new(Length::Units(5), Length::Units(0)))
-                .push(delete_button.map(Message::Interaction))
-                .width(Length::Fill);
-            let column = Column::new()
-                .push(author_title_container)
-                .push(Space::new(Length::Units(0), Length::Units(3)))
-                .push(author_text)
-                .push(Space::new(Length::Units(0), Length::Units(15)))
-                .push(notes_title_container)
-                .push(Space::new(Length::Units(0), Length::Units(3)))
-                .push(notes_text)
-                .push(Space::new(Length::Units(0), Length::Units(15)))
-                .push(release_channel_title_container)
-                .push(Space::new(Length::Units(0), Length::Units(3)))
-                .push(test_row)
-                .push(space)
-                .push(button_row)
-                .push(bottom_space);
-            let details_container = Container::new(column)
-                .width(Length::Fill)
-                .padding(20)
-                .style(style::FadedNormalForegroundContainer(color_palette));
-
-            let row = Row::new()
-                .push(left_spacer)
-                .push(details_container)
-                .push(Space::new(
-                    Length::Units(DEFAULT_PADDING + 5),
-                    Length::Units(0),
-                ))
-                .spacing(1);
-
-            addon_column = addon_column
-                .push(Space::new(Length::FillPortion(1), Length::Units(1)))
-                .push(row);
-        }
-    }
-
-    Container::new(addon_column)
-        .width(Length::Fill)
-        .style(style::Row(color_palette))
-}
-
 fn row_title<T: PartialEq>(
     column_key: T,
     previous_column_key: Option<T>,
